Add step to create hybrid logic app
